refactor: bottom nav 색상변경 및 사과 이미지 -> 스낵 이미지
