+++ src/components/BottomNav/BottomNav.tsx
@@ -12,7 +12,7 @@ export const BottomNav = () => {
 
   return (
     <div className="fixed bottom-0 z-[999] flex w-full justify-center">
-      <div className="flex h-16 w-full max-w-3xl items-center justify-around rounded-t-2xl bg-white text-gray-500 shadow-navigation">
+      <div className="flex h-16 w-full max-w-3xl items-center justify-around rounded-t-2xl bg-primary-light text-gray-500 shadow-navigation">
         <RouterLink
           to={PATH.SNACK_GAME}
           className={'flex-col items-center justify-between'}
