+++ src/pages/games/SnackGame/game/model/snackGame/snack.ts
@@ -125,7 +125,6 @@ abstract class Snack {
     ctx.font = `${this.radius}px Dovemayo_gothic`;
     ctx.fillStyle = '#ffffff';
     ctx.textAlign = 'center';
-    ctx.textBaseline = 'middle';
     ctx.fillText(
       String(this.snackNumber),
       this.position.x + this.radius,
